requirements.txt
pystray~=0.19.5
pillow~=11.2.1
flet~=0.28.3
keyboard~=0.13.5
torch~=2.7.1
sounddevice~=0.5.2
numpy~=2.2.6
PyAutoGUI~=0.9.54
requests~=2.32.4
DateTime~=5.5
vosk~=0.3.45
psutil~=7.0.0
pywinauto~=0.6.9
pycaw~=20240210
comtypes~=1.4.11
beautifulsoup4~=4.13.4
python-dotenv~=1.1.1